+++ .gitignore
@@ -1,3 +1,3 @@
-output/
-font/
+tests/output/
+tests/fonts/
 *.exe
