+++ go.mod
@@ -1,11 +1,17 @@
-module qr-gen
+module github.com/TOsmanov/qr-gen
 
 go 1.21.5
 
-require github.com/skip2/go-qrcode v0.0.0-20200617195104-da1b6568686e
+require (
+	github.com/fogleman/gg v1.3.0
+	github.com/skip2/go-qrcode v0.0.0-20200617195104-da1b6568686e
+)
 
 require (
-	github.com/fogleman/gg v1.3.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
-	golang.org/x/image v0.15.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/testify v1.9.0 // indirect
+	golang.org/x/image v0.16.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
